+++ examples/dockermodelrunner/toolcall.php
@@ -1,0 +1,34 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+use Symfony\AI\Agent\Agent;
+use Symfony\AI\Agent\Toolbox\AgentProcessor;
+use Symfony\AI\Agent\Toolbox\Tool\Wikipedia;
+use Symfony\AI\Agent\Toolbox\Toolbox;
+use Symfony\AI\Platform\Bridge\DockerModelRunner\Completions;
+use Symfony\AI\Platform\Bridge\DockerModelRunner\PlatformFactory;
+use Symfony\AI\Platform\Message\Message;
+use Symfony\AI\Platform\Message\MessageBag;
+
+require_once dirname(__DIR__).'/bootstrap.php';
+
+$platform = PlatformFactory::create(env('DOCKER_MODEL_RUNNER_HOST_URL'), http_client());
+$model = new Completions(Completions::GEMMA_3_N);
+
+$wikipedia = new Wikipedia(http_client());
+$toolbox = new Toolbox([$wikipedia]);
+$processor = new AgentProcessor($toolbox);
+$agent = new Agent($platform, $model, [$processor], [$processor], logger: logger());
+
+$messages = new MessageBag(Message::ofUser('Who is the actual Prime Minister of France?'));
+$result = $agent->call($messages);
+
+echo $result->getContent().\PHP_EOL;